Address comments and refactor
