+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/counters/MeanCounterUpdateAggregator.java
@@ -15,25 +15,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.dataflow.worker;
+package org.apache.beam.runners.dataflow.worker.counters;
 
 import static org.apache.beam.runners.dataflow.worker.counters.DataflowCounterUpdateExtractor.longToSplitInt;
 import static org.apache.beam.runners.dataflow.worker.counters.DataflowCounterUpdateExtractor.splitIntToLong;
 
 import com.google.api.services.dataflow.model.CounterUpdate;
+import com.google.api.services.dataflow.model.IntegerMean;
 import java.util.List;
-import org.apache.beam.runners.dataflow.worker.MetricsToCounterUpdateConverter.Kind;
 
-public class SumCounterUpdateAggregator implements CounterUpdateAggregator {
+public class MeanCounterUpdateAggregator implements CounterUpdateAggregator {
 
   @Override
   public CounterUpdate aggregate(List<CounterUpdate> counterUpdates) {
     if (counterUpdates == null || counterUpdates.isEmpty()) {
       return null;
     }
-    if (counterUpdates.stream().anyMatch(c -> c.getInteger() == null)) {
+    if (counterUpdates.stream().anyMatch(c -> c.getIntegerMean() == null)) {
       throw new UnsupportedOperationException(
-          "Aggregating SUM counter updates over non-integer type is not implemented.");
+          "Aggregating MEAN counter updates over non-integerMean type is not implemented.");
     }
 
     CounterUpdate initial = counterUpdates.remove(0);
@@ -41,13 +41,15 @@ public class SumCounterUpdateAggregator implements CounterUpdateAggregator {
         .reduce(
             initial,
             (first, second) ->
-                first.setInteger(
-                    longToSplitInt(
-                        splitIntToLong(first.getInteger()) + splitIntToLong(second.getInteger()))));
-  }
-
-  @Override
-  public Kind getKind() {
-    return Kind.SUM;
+                first.setIntegerMean(
+                    new IntegerMean()
+                        .setCount(
+                            longToSplitInt(
+                                splitIntToLong(first.getIntegerMean().getCount())
+                                    + splitIntToLong(second.getIntegerMean().getCount())))
+                        .setSum(
+                            longToSplitInt(
+                                splitIntToLong(first.getIntegerMean().getSum())
+                                    + splitIntToLong(second.getIntegerMean().getSum())))));
   }
 }
