+++ runners/google-cloud-dataflow-java/worker/src/test/java/org/apache/beam/runners/dataflow/worker/counters/DistributionCounterUpdateAggregatorTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.dataflow.worker;
+package org.apache.beam.runners.dataflow.worker.counters;
 
 import static org.apache.beam.runners.dataflow.worker.counters.DataflowCounterUpdateExtractor.longToSplitInt;
 import static org.apache.beam.runners.dataflow.worker.counters.DataflowCounterUpdateExtractor.splitIntToLong;
@@ -24,43 +24,49 @@ import static org.junit.Assert.assertEquals;
 import com.google.api.services.dataflow.model.CounterMetadata;
 import com.google.api.services.dataflow.model.CounterStructuredNameAndMetadata;
 import com.google.api.services.dataflow.model.CounterUpdate;
-import com.google.api.services.dataflow.model.IntegerMean;
+import com.google.api.services.dataflow.model.DistributionUpdate;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.beam.runners.dataflow.worker.MetricsToCounterUpdateConverter.Kind;
 import org.junit.Before;
 import org.junit.Test;
 
-public class MeanCounterUpdateAggregatorTest {
+public class DistributionCounterUpdateAggregatorTest {
 
   private List<CounterUpdate> counterUpdates;
-  private MeanCounterUpdateAggregator aggregator;
+  private DistributionCounterUpdateAggregator aggregator;
 
   @Before
   public void setUp() {
     counterUpdates = new ArrayList<>();
-    aggregator = new MeanCounterUpdateAggregator();
+    aggregator = new DistributionCounterUpdateAggregator();
     for (int i = 0; i < 10; i++) {
       counterUpdates.add(
           new CounterUpdate()
               .setStructuredNameAndMetadata(
                   new CounterStructuredNameAndMetadata()
                       .setMetadata(new CounterMetadata().setKind(Kind.MEAN.toString())))
-              .setIntegerMean(
-                  new IntegerMean().setSum(longToSplitInt((long) i)).setCount(longToSplitInt(1L))));
+              .setDistribution(
+                  new DistributionUpdate()
+                      .setSum(longToSplitInt((long) i))
+                      .setMax(longToSplitInt((long) i))
+                      .setMin(longToSplitInt((long) i))
+                      .setCount(longToSplitInt((long) 1))));
     }
   }
 
   @Test
   public void testAggregate() {
     CounterUpdate combined = aggregator.aggregate(counterUpdates);
-    assertEquals(45L, splitIntToLong(combined.getIntegerMean().getSum()));
-    assertEquals(10L, splitIntToLong(combined.getIntegerMean().getCount()));
+    assertEquals(45L, splitIntToLong(combined.getDistribution().getSum()));
+    assertEquals(10L, splitIntToLong(combined.getDistribution().getCount()));
+    assertEquals(9L, splitIntToLong(combined.getDistribution().getMax()));
+    assertEquals(0L, splitIntToLong(combined.getDistribution().getMin()));
   }
 
   @Test(expected = UnsupportedOperationException.class)
-  public void testAggregateWithNullIntegerMean() {
-    counterUpdates.get(0).setIntegerMean(null);
+  public void testAggregateWithNullIntegerDistribution() {
+    counterUpdates.get(0).setDistribution(null);
     aggregator.aggregate(counterUpdates);
   }
 }
